requirements.txt
asgiref==3.8.1
cloudinary==1.41.0
Django==3.2.4
django-cloudinary-storage==0.3.0
djangorestframework==3.12.4
pillow==11.0.0
pytz==2024.2
sqlparse==0.5.1
